requirements.txt
streamlit==1.29.0
snowflake-connector-python==3.6.0
pandas==2.1.4
plotly==5.17.0
# Google OAuth dependencies removed - using simple authentication
# streamlit-authenticator==0.2.3  # Removed - using custom simple auth
python-dotenv==1.0.0
altair==5.2.0
numpy==1.24.3
openpyxl==3.1.2
networkx>=2.8.8
scikit-learn>=1.3.0
scipy>=1.11.0
datetime 